Move pkg/duck to pkg/duck/v1alpha1.
