+++ pkg/apis/events/v1alpha1/cloudbuildsource_types.go
@@ -17,7 +17,7 @@ import (
 	"fmt"
 
 	duckv1alpha1 "github.com/google/knative-gcp/pkg/apis/duck/v1alpha1"
-	kngcpduck "github.com/google/knative-gcp/pkg/duck"
+	kngcpduck "github.com/google/knative-gcp/pkg/duck/v1alpha1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"knative.dev/pkg/apis/duck"
 	"knative.dev/pkg/kmeta"
