+++ pkg/duck/v1alpha1/pubsubable.go
